Added functionality and tests for GET articles by ID and separated error handling into a different file for readability
